[Move] Overlay.css: Contacts -> Base
